srv-admin/requirements.txt
bandit
black
django-environ==0.4.5
django-filter==2.4.0
django-grappelli==2.14.3
django-imagekit==4.0.2
django-import-export==2.5.0
django-modeltranslation==0.16.2
django-push-notifications==2.0.0
Django==3.1.5
djangorestframework-jwt==1.11.0
djangorestframework==3.12.2
drf-extra-fields==3.1.1
easy-thumbnails==2.7.1
ipdb==0.13.4
ipython==7.19.0
mock==4.0.3
pdfkit==0.6.1
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.6
pytest-django==4.1.0
pytest-mock==3.5.0
pytest==6.2.1
pyyaml
requests_pkcs12
swapper==1.1.2.post1
Unipath==1.1
uritemplate
xmltodict==0.12.0
django-cors-headers==3.7.0
